requirements.txt
asgiref==3.2.3
Django==3.0.7
Pillow==7.1.0
pytz==2019.3
sqlparse==0.3.0
django-colorfield==0.4.1
django-crispy-forms==1.11.2
